requirements_dev.txt
alembic==1.9.1
anyio==3.6.2
attrs==22.2.0
black==22.12.0
cachelib==0.9.0
certifi==2022.12.7
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.0
filelock==3.9.0
Flask==2.2.2
Flask-Caching==2.0.2
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
identify==2.5.13
idna==3.4
iniconfig==2.0.0
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==23.0
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==2.21.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
PyYAML==6.0
rfc3986==1.5.0
sendgrid==6.9.7
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starkbank-ecdsa==2.2.0
tomli==2.0.1
typing_extensions==4.4.0
virtualenv==20.17.1
Werkzeug==2.2.2
WTForms==3.0.1
